refactor(bootcamp): de-duplicate shim-log/stripMdExt helpers + harness spawn scaffold
Eliminates the bootcamp-introduced duplication that pushed the two
duplication health gates over budget, without touching any budget:

- Hoist stripMdExt + loadShimLog into scenarios/_helpers.ts; b1/b2/b4/b7
  now import the shared copies instead of carrying near-identical locals.
- Extract the identical spawn → transcript → SIGKILL → parse shell from
  the claude and codex drivers into drivers/_harnessProcess.ts
  (runHarnessProcess); both runScenario impls become thin specs of
  {command, args, parse}.

Gates after refactor (measures, tier-1-health):
- semantic-duplication >=0.7 pairs: 653 -> 645 (budget 645)
- recoverable-LOC mass: 2638 -> 2570 (budget 2586)
- full tier-1-health suite: 248 passed / 1 skipped / 0 failed

Behaviour preserved: all 134 bootcamp tests pass, including the driver
and B1/B2/B4/B7 scenario suites that exercise the refactored code.
