+++ packages/systems/voicetree-bootcamp/src/drivers/claude.ts
@@ -2,8 +2,9 @@
  * Claude Code harness driver.
  *
  * Pure parseClaudeStream interprets the --output-format=stream-json NDJSON;
- * impure runScenario spawns `claude`, captures stdout to a transcript file,
- * and folds in driver-measured wall-clock time.
+ * impure runScenario delegates to runHarnessProcess to spawn `claude`,
+ * capture stdout to a transcript file, and fold in driver-measured
+ * wall-clock time.
  *
  * Token strategy (corrected after probe 3): Claude Code emits one
  * type:"assistant" event per content block within a single message.id, and
@@ -13,11 +14,8 @@
  * result line, fall back to deduping assistant events by message.id and
  * summing one entry per unique message.
  */
-import {spawn} from 'node:child_process'
-import {createWriteStream} from 'node:fs'
-import {readFile} from 'node:fs/promises'
-import {join} from 'node:path'
 import type {Effort, HarnessDriver, RunTelemetry} from '../types.ts'
+import {runHarnessProcess} from './_harnessProcess.ts'
 
 type Usage = {
     input_tokens?: number
@@ -155,62 +153,23 @@ function isResultEvent(v: Record<string, unknown>): v is ResultEvent {
 export const claudeCodeDriver: HarnessDriver = {
     name: 'claude',
     models: ['haiku', 'sonnet', 'opus'],
-    runScenario: async (opts) => {
-        const transcriptPath = join(opts.artifactDir, 'transcript.jsonl')
-        const args = [
-            '--print',
-            '--model',
-            opts.model,
-            '--effort',
-            opts.effort satisfies Effort,
-            '--permission-mode',
-            'bypassPermissions',
-            '--output-format',
-            'stream-json',
-            '--verbose',
-            '--',
-            opts.prompt,
-        ]
-
-        const start = Date.now()
-        const child = spawn('claude', args, {
-            cwd: opts.cwd,
-            env: {...opts.env},
-            stdio: ['ignore', 'pipe', 'pipe'],
-        })
-
-        const transcriptStream = createWriteStream(transcriptPath)
-        child.stdout.pipe(transcriptStream)
-        child.stderr.resume()
-
-        const timeoutHandle = setTimeout(() => {
-            child.kill('SIGKILL')
-        }, opts.timeoutMs)
-
-        const exitInfo = await new Promise<{
-            code: number | null
-            signal: NodeJS.Signals | null
-        }>((resolve) => {
-            child.on('close', (code, signal) => {
-                clearTimeout(timeoutHandle)
-                resolve({code, signal})
-            })
-        })
-
-        const wallClockMs = Date.now() - start
-
-        await new Promise<void>((resolve) => {
-            if (transcriptStream.closed) resolve()
-            else transcriptStream.end(() => resolve())
-        })
-
-        const ndjson = await readFile(transcriptPath, 'utf8')
-        const {telemetry} = parseClaudeStream(ndjson)
-
-        return {
-            transcriptPath,
-            telemetry: {...telemetry, wallClockMs},
-            exitInfo,
-        }
-    },
+    runScenario: (opts) =>
+        runHarnessProcess(opts, {
+            command: 'claude',
+            args: [
+                '--print',
+                '--model',
+                opts.model,
+                '--effort',
+                opts.effort satisfies Effort,
+                '--permission-mode',
+                'bypassPermissions',
+                '--output-format',
+                'stream-json',
+                '--verbose',
+                '--',
+                opts.prompt,
+            ],
+            parse: parseClaudeStream,
+        }),
 }
